backend/requirements.txt
algofi-py-sdk @ git+https://github.com/Algofiorg/algofi-py-sdk@4c23e6064a0f33ccb538421f2c6db412d3d4e9ea
anyio==3.5.0
asgiref==3.5.0
attrs==19.3.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
durationpy==0.5
ecdsa==0.17.0
fastapi==0.75.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
msgpack==1.0.3
mypy-extensions==0.4.3
passlib==1.7.4
py-algorand-sdk==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.14.1
pydantic==1.9.0
pymongo==4.1.1
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.0.3
starlette==0.17.1
typing-extensions==3.7.4.3
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.0
zipp==3.8.0
